+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased Changes
 * Added `Arena::with_capacity` for preallocating space.
+* Added `Arena::len`, `Arena::capacity`, and `Arena::is_empty`.
 
 ## 0.1.0 (2020-09-02)
 * Initial release